un-skipping gather test

+++ src/lazy/dsl.rs
@@ -330,7 +330,7 @@ impl JsExpr {
     }
     #[napi(catch_unwind)]
     pub fn gather(&self, idx: &JsExpr) -> JsExpr {
-        self.clone().inner.gather(idx.inner.clone()).into()
+        self.clone().inner.gather(idx.inner.clone().cast(DataType::Int64)).into()
     }
     #[napi(catch_unwind)]
     pub fn sort_by(&self, by: Vec<&JsExpr>, reverse: Vec<bool>) -> JsExpr {
